+++ wsgidav/server/server_sample.py
@@ -7,7 +7,7 @@ Simple example how to a run WsgiDAV in a 3rd-party WSGI server.
 """
 from cheroot import wsgi
 
-from wsgidav import __version__, util
+from wsgidav import util
 from wsgidav.fs_dav_provider import FilesystemProvider
 from wsgidav.wsgidav_app import WsgiDAVApp
 
@@ -36,7 +36,7 @@ def main():
 
     # For an example, use cheroot:
     version = (
-        f"WsgiDAV/{__version__} {wsgi.Server.version} Python/{util.PYTHON_VERSION}"
+        f"{util.public_wsgidav_info} {wsgi.Server.version} Python/{util.PYTHON_VERSION}"
     )
 
     server = wsgi.Server(